deploy/requirements.txt
fabric>=3.0
pyyaml
requests
